refactored routes;removed routes main.tsx to separate file
