+++ src/lib/auth.ts
@@ -10,3 +10,14 @@ export async function verifySession(userId: string, secret: string) {
   const data = await account.updateMagicURLSession(userId, secret);
   return data;
 }
+
+export async function getCurrentSession() {
+  const session = await account.getSession('current');
+  return {
+    session
+  }
+}
+
+export async function deleteCurrentSession() {
+  await account.deleteSession('current');
+}
